Wireless Bluetooth Connections
The objective of this lab is to interface a Bluetooth communications module to a RoboRed microcontroller. The Bluetooth connection is used to the send a command via a smartphone to control a
beeper module and to receive joystick X and Y output data from the micro controller to the
smartphone.